Blog Completed
Blog is completed, I don't think he wants a fleshed out blog system. I use what we learned from the book to do this so, i am assuming that is what he wants. If someone knows how to get the CRUD to work that would be awesome, I gotta research this, if we need it. I am thinking about just removing the links to the CRUD.